change annotate parameter of mauve. min_cover_ratio is float now.

+++ src/data.ml
@@ -70,7 +70,7 @@ type re_meth_t = [ `Locus_Breakpoint of int |
 type median_solver_t = [ `MGR | `Vinh | `Albert | `Siepel | `BBTSP | `COALESTSP |
 `ChainedLK | `SimpleLK ]
 
-type annotate_tool_t = [ `Mauve of (float*int*int) | `Default of (int*int*int) ]
+type annotate_tool_t = [ `Mauve of (float*float*int) | `Default of (int*int*int) ]
 
 type dyna_pam_t = {
